(svn r8661) -Fix: [depot windows] Enlarge the blocks in the depot window if a sprite is too big to fit (ships and aircraft only) -Codechange: depot windows default now makes a 1x1 matrix and resize the window by block size*number of blocks    This will make less vehicle type specific code and easier to read resizing based on sprite sizes

+++ src/vehicle_gui.h
@@ -64,7 +64,7 @@ static inline void DrawVehicleImage(const Vehicle *v, int x, int y, int count, i
 	}
 }
 
-static inline byte GetVehicleListHeight(byte type)
+static inline uint GetVehicleListHeight(byte type)
 {
 	return (type == VEH_Train || type == VEH_Road) ? 14 : 24;
 }
